+++ packages/aware-components/src/components/Table/Table.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { DEVELOPMENT } from '../../constants';
+import { useTable } from '../../context/table/actions';
 import TableProvider from '../../context/table/provider';
 import { warn } from '../../helper/consoleWarn';
 import { a11yChecks } from '../../utils/a11y';
@@ -14,19 +15,18 @@ interface Props
 
 export function Development(props: Props) {
   const { a11y = true, children, ...rest } = props;
+  const { header, caption } = useTable();
 
-  if (a11y) a11yChecks.table(props)?.forEach(warn);
+  if (a11y) a11yChecks?.table?.(props, header, caption)?.forEach(warn);
 
-  return (
-    <TableProvider>
-      <table {...rest}>{children}</table>
-    </TableProvider>
-  );
+  return <table {...rest}>{children}</table>;
 }
 
 export const Table = (props: Props) =>
   DEVELOPMENT ? (
-    <Development {...props} />
+    <TableProvider>
+      <Development {...props} />
+    </TableProvider>
   ) : (
     <table {...props}>{props.children}</table>
   );