+++ packages/aware-components/src/utils/a11y/table/checks/checkColHeader.ts
@@ -1,5 +1,4 @@
-import { ReactElement } from 'react';
 import { messages } from '../../../messages';
 
-export const checkColHeader = (colHeaders: ReactElement[]) =>
-  !colHeaders.length ? messages.table.col : null;
+export const checkColHeader = (hasColumnHeader: boolean) =>
+  !hasColumnHeader ? messages.table.col : null;