+++ packages/aware-components/src/utils/a11y/table/checks/checkRowHeader.ts
@@ -2,12 +2,5 @@ import { ReactElement } from 'react';
 import { validateScopeAttribute } from '../../../../helper/tables';
 import { messages } from '../../../messages';
 
-export const checkRowHeader = (
-  colHeaders: ReactElement[],
-  rowHeaders: ReactElement[]
-) =>
-  colHeaders.length &&
-  rowHeaders.length &&
-  !validateScopeAttribute([...colHeaders, ...rowHeaders])
-    ? messages.table.row
-    : null;
+export const checkRowHeader = (headers: ReactElement[]): string | null =>
+  !validateScopeAttribute(headers) ? messages.table.row : null;