+++ packages/aware-components/src/utils/a11y/td/index.ts
@@ -1,5 +1,5 @@
 import { Scope } from '../../../context/table/types';
-import { isMultiHeaderTable } from '../../../helper/tables';
+import { isMultiHeadingTable } from '../../../helper/tables';
 import { messages } from '../../messages';
 
 type Props = React.DetailedHTMLProps<
@@ -11,7 +11,7 @@ const warning = ({ children }: Props) =>
   `${messages.td.multi}"${children || ''}"`;
 
 export const tdChecks = (props: Props, header: Scope[]): string[] => [
-  ...(isMultiHeaderTable(header) &&
+  ...(isMultiHeadingTable(header) &&
   (props.children || Object.keys(props).length) &&
   !props.headers
     ? [warning(props)]