+++ packages/playground/src/components/Tables/Description.tsx
@@ -1,0 +1,5 @@
+import { Caption } from 'aware-components';
+
+export function Description() {
+  return <Caption>Delivery slots:</Caption>;
+}